Incremental work to support dynamic visualization selection by the user. Added a new ListOverlay type (not quite done yet).

+++ src/musikbox/cursespp/MessageQueue.cpp
@@ -138,7 +138,7 @@ void MessageQueue::Post(IMessagePtr message, int64 delayMs) {
     this->queue.insert(curr, m);
 }
 
-void MessageQueue::Debounce(IMessagePtr message, int delayMs) {
+void MessageQueue::Debounce(IMessagePtr message, int64 delayMs) {
     Remove(message->Target(), message->Type());
     Post(message, delayMs);
 }
